Update to use proper extension methods
Core no longer fires an event as Components are extendable by
default, so this updates the code to reflect that change.

In addition, it turns out methods *can* be added to extendable
classes, meaning there's no need for the old registerFilter()
workaround. Instead, we'll add the scope directly to the model.